Add ability to disable pkcs7 key generation with hiera-eyaml
Without this change hiera-eyaml generates pkcs7 keys automatically if
the eyaml backend is enabled. This adds a parameter called create_keys
to allow disabling key creation. The parameter defaults to true, so the
current module behavior is preserved. If set to false you are able to
manage the keys outside of this module. Adding a default owner and group
for file resources in eyaml.pp to reduce duplication.