Fixing 2015.2 handling
This PE stuff is confusing.

This should fix the hiera-eyaml installation on PE 2015.2. It also fixes
a bug with versioncmp() trying to compare $::pe_server_version when it
is undef to determine if it's on PE or FOSS puppetserver.

It also updates the formatting of hiera.yaml to be a bit more
consistent across puppet versions (thanks zaml for being not like yaml).

This adds beaker tests that should work on foss puppet 3, PE 3, and
PE 2015.2+puppet-agent packaging.