Ugly hack to workaround change in rspec-puppet
rspec-puppet previously defaulted `confdir` to `/etc/puppet`.
It now sets it to `/dev/null`.

Since puppet 4 it's been `/etc/puppetlabs/puppet`, but if we try to set
it to this ourselves in the `RSpec.configure` block in spec_helper.rb it
only creates more problems.  rspec-puppet (or maybe puppet itself) gets
upset when it tries to create the parent directory (`/etc/puppetlabs`)
and gets a `permission denied`.  Furthermore, we'd have to make changes
to modulesync_config to accommodate these changes.

It's ugly, but for now, the easiest thing to do is just pretend
`/dev/null` is a valid `configdir`.  Sigh.