Add test for eyaml_gpg and some other minor checks
I found some checks missing from rake runs, and fixed
them. Doing this I mangled hiera_spec.rb to achieve
consistent naming of keysdir. When I started I used
default and that caused 3 different namings of the
same directory. That seems silly.

The reason it ended up being /dev/null/keys is that
theam YAML eyaml loader sets confdir to "/dev/null"
The string "/dev/null" is not mentioned anywhere in
the code. So I went with it and changed keysdir to
/dev/null/keys instead of /etc/keys. This is a small
change, that increases the coverage percentage when
I check some file in the eyaml_gpg_spec.rb.

Also added a check for the deep_merge class, just
becasue it was easy and on my way. I am getting the
hang of this now - but if you would ask me what I
have done and how, I would have to do answer in a
Mr. Miagi-style.

  Don't know. Firsta time.